moving to tensor file
